requirements.txt
# 核心依赖
DrissionPage==4.0.4.22
PyYAML>=6.0
loguru>=0.7.0
python-dotenv>=1.0.0

# AI 提供商（二选一或都安装）
# Google Gemini（推荐）
google-generativeai>=0.3.0

# OpenAI（可选）
openai>=1.0.0

# 文档处理
pdfplumber>=0.10.0
docxtpl>=0.16.0
python-docx>=1.1.0

# 定时任务
APScheduler>=3.10.0

# 网络请求
requests>=2.31.0

# 数据库
# SQLite 是 Python 内置的，无需额外安装

# 其他工具
pytz>=2023.3
